Fix non-existing tasks in Lage Info if transient tasks are noop
Tasks defined in the Lage pipeline configuration ends up as noop tasks in each
package and is used as a path to resolve ancestors when depending on parent
tasks.

Say a task "build" depends on "^build", which means it depends on all its
parents "build" tasks to be run before itself, then if the direct dependency
doesn't have a verbatim "build" task it'll still get a noop task that it can
depend on, which again can have dependencies with real build taks that then gets
exectured.

Problem is `lage info` prints out these noop tasks as they existed. This change
resolves the tasks recursively, so if a task depends on a noop task we rewrite
the dependencies to rather point to first real task up the line.