Fix issue with noop tasks that depend on other noop tasks
If you had a runnable task that depended on a noop task that in tern depended on a noop task, that depended on a runnable task, the dependency between the two runnable tasks would not be preserved.  The logic that prunes out noop tasks in the dependency tree did not account for multiple levels of removal.